feat(api-server): /agents /skills /tools /teams /projects /tasks routers + RLS-aware helpers
Six tenant-scoped REST resources wired through a shared _helpers
module (require_tenant_id, get_writable_or_404, apply_partial_update,
soft_delete). Built-in rows return 404 on PUT/DELETE/POST to avoid
leaking which IDs are platform-owned.

Each resource has its own Pydantic schema module with separate
Create / Update / Response models. agents.py includes the fork,
diff and merge endpoints (migrations land in the next commit).
AgentResponse uses Pydantic alias `model_config` <-> `llm_config`
to dodge the BaseModel reserved name; constructed via model_validate
so mypy stays happy.

Also fixes Project.budget_amount typing (float -> Decimal) so the
schema's Decimal field aligns with the SQLAlchemy Numeric column.

Covers tasks 01_05..08.

+++ apps/api-server/src/api_server/db/domain.py
@@ -27,6 +27,7 @@ from __future__ import annotations
 
 import enum
 from datetime import datetime
+from decimal import Decimal
 from typing import Any
 from uuid import UUID
 
@@ -497,7 +498,9 @@ class Project(Base, UUIDPrimaryKeyMixin, TenantScopedMixin, TimestampMixin, Soft
     secrets_vault_id: Mapped[UUID | None] = mapped_column(PG_UUID(as_uuid=True), nullable=True)
 
     # --- Budget (see spec §28.7 for tenant-vs-project budget interaction) ---
-    budget_amount: Mapped[float | None] = mapped_column(
+    # Numeric(14,2) maps to Decimal in Python so the schema's Decimal
+    # type stays consistent end-to-end (no float rounding on currency).
+    budget_amount: Mapped[Decimal | None] = mapped_column(
         Numeric(precision=14, scale=2), nullable=True
     )
     budget_currency: Mapped[str | None] = mapped_column(String(3), nullable=True)
